requirements.txt
code128==0.3
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
sendgrid==6.11.0
six==1.16.0
starkbank-ecdsa==2.2.0
tzdata==2024.1
